move transactions state files
