--- src/redux/selectors/updates/index.ts
@@ -1,1 +1,0 @@
-export * from './transactions';
